requirements.txt
Django>=4.2.0,<5.0.0
python-decouple>=3.8
whitenoise>=6.5.0
psycopg2-binary>=2.9.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2021.1.0
django-imagefit>=0.7.2
django-resized>=1.0.3
boto3==1.26.59
django-storages==1.13.2